+++ lib/class_admin_tools.php
@@ -89,6 +89,42 @@ if( !class_exists( 'class_admin_tools' ) ) {
       $data[ 'Developer' ][ ] = '<p>' . __( '<b>Search Input:</b> Select and input type and enter comma-separated values that you would like to be used in property search, on the front-end.', ud_get_wp_property()->domain ) . '</p>';
       $data[ 'Developer' ][ ] = '<p>' . __( '<b>Data Entry:</b> Enter comma-separated values that you would like to use on the back-end when editing properties.', ud_get_wp_property()->domain ) . '</p>';
 
+
+      // Google map api key
+      $data[ 'Google Map API key' ][ ] = '<h3>' . __( 'Setting up Google Map API keys', ud_get_wp_property()->domain ) . '</h3>';
+      $data[ 'Google Map API key' ][ ] = '<p>' . __( "If your client application does not use OAuth 2.0, then it must include an API key when it calls an API that's enabled within a Google Cloud Platform project. The application passes this key into all API requests as a key=API_key parameter.", ud_get_wp_property()->domain ) . '</p>';
+      $data[ 'Google Map API key' ][ ] = '<p>' . __( "To create your application's API key:", ud_get_wp_property()->domain ) . '</p>';
+
+      $data[ 'Google Map API key' ][ ] = '<ol>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . sprintf(__( "Go to the <a href='%s'>API Console</a>.", ud_get_wp_property()->domain ), 'https://console.developers.google.com/') . '</li>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "From the projects list, select a project or create a new one.", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . sprintf(__( "If the API Manager page isn't already open, open the menu %s and select <b>API Manager</b>.", ud_get_wp_property()->domain ), "&nbsp;<img src='data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAA8AAAAKCAYAAABrGwT5AAAABHNCSVQICAgIfAhkiAAAADZJREFUKJFjrKio+M9AJmD59+/fGXI1DxxgLCsre0euZiZquoR+gLGsrOw0uZpZmJiYTMjVDAAtHArypQrPMgAAAABJRU5ErkJggg==' height='10'>&nbsp;") . '</li>';
+
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "From Dashboard click <b>Enable API</b> button at top or click on <b>Library</b> from left menu.", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "Click <b>Google Maps JavaScript API</b> and <b>Google Maps Geolocation API </b> from <b>Google Maps APIs</b> one by one.", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "To enable the api click <b>Enable</b>. Do this for both <b>Google Maps JavaScript API</b> and <b>Google Maps Geolocation API </b>,", ud_get_wp_property()->domain ) . '</li>';
+
+      
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "On the left, choose <b>Credentials</b>.", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "Click <b>Create credentials</b> and then select API key.", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "Copy API key from <b>Your API key</b> field.", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '</ol>';
+
+      $data[ 'Google Map API key' ][ ] = '<h4>Optional but highly recommended (To to prevent unauthorized use in production):</h4>';
+
+      $data[ 'Google Map API key' ][ ] = '<ol start="7">';
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "Click <b>Restrict key</b> button.", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "Optional: You can give your API key a name in name field. Useful when you have lots of API keys.", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "Select <b>HTTP referrers (web sites)</b> for Browser Key and <b>IP addresses (web servers, cron jobs, etc.)</b> for Server Key.", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "Add your web site address (for browser key) or server IP addresses (for server key) to next input.", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "Click save.", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '<li>' . __( "You have do whole process couple of time. Once for <b>Browser key</b> and once for <b>Server key</b>", ud_get_wp_property()->domain ) . '</li>';
+      $data[ 'Google Map API key' ][ ] = '</ol>';
+
+      $data[ 'Google Map API key' ][ ] = '<p>' . sprintf(__( 'For more info <a target="_blank" href="%s">click here</a>', ud_get_wp_property()->domain ), "https://developers.google.com/maps/documentation/geocoding/get-api-key") . '</p>';
+      $data[ 'Google Map API key' ][ ] = '<p>' . sprintf(__( 'You can also read <a target="_blank" href="%s">Best practices for securely using API keys</a>', ud_get_wp_property()->domain ), "https://support.google.com/googleapi/answer/6310037") . '</p>';
+
+
       return $data;
 
     }