+++ static/views/admin/settings.php
@@ -209,13 +209,22 @@ if ( get_option( 'permalink_structure' ) == '' ) {
     <?php endif; ?>
 
     <tr>
-      <th><?php printf( __( 'Geolocation options', ud_get_wp_property()->domain ), WPP_F::property_label() ); ?></th>
+      <th><?php printf( __( 'Automatic Geolocation', ud_get_wp_property()->domain ), WPP_F::property_label() ); ?></th>
       <td>
         <ul>
+          <li><?php _e( 'Attribute to use for physical addresses:', ud_get_wp_property('domain') ); ?><?php echo WPP_F::draw_attribute_dropdown( "name=wpp_settings[configuration][address_attribute]&selected={$wp_properties[ 'configuration' ]['address_attribute']}" ); ?></li>
           <li><?php _e( 'Localize addresses in:', ud_get_wp_property('domain') ); ?> <?php echo WPP_F::draw_localization_dropdown( "name=wpp_settings[configuration][google_maps_localization]&selected={$wp_properties[ 'configuration' ]['google_maps_localization']}" ); ?></li>
           <li class="google-maps-api-section" data-feature-since="2.0.3">
-            <?php printf(__( 'Google Maps API (optional):', ud_get_wp_property('domain') ) ); ?> <?php echo WPP_F::input( "name=wpp_settings[configuration][google_maps_api]", ud_get_wp_property( 'configuration.google_maps_api' ) ); ?>
-            <br/><span class="description"><?php printf( __( 'Note, Google Maps has its own limit of usage. You can provide Google Maps API license ( key ) above to increase limit. See more details %shere%s.', ud_get_wp_property('domain') ), '<a href="https://developers.google.com/maps/documentation/javascript/usage#usage_limits" target="_blank">', '</a>' ); ?></span>
+            <?php printf(__( 'Google Maps API (Browser Key):', ud_get_wp_property('domain') ) ); ?>
+            <?php echo WPP_F::input( "name=wpp_settings[configuration][google_maps_api]", ud_get_wp_property( 'configuration.google_maps_api' ) ); ?>
+
+            <br/><span class="description"><?php printf( __( 'Note, Google Maps has its own limit of usage. You need to provide Google Maps API license ( browser key ) above to increase limit. See more details in %shelp tab%s.', ud_get_wp_property('domain') ), '<a href="#tab-link-google-map-api-key" class="open-help-tab">', '</a>' ); ?></span>
+          </li>
+          <li class="google-maps-api-section" data-feature-since="2.0.3">
+            <?php printf(__( 'Google Maps API (Server Key):', ud_get_wp_property('domain') ) ); ?>
+            <?php echo WPP_F::input( "name=wpp_settings[configuration][google_maps_api_server]", ud_get_wp_property( 'configuration.google_maps_api_server' ) ); ?>
+
+            <br/><span class="description"><?php printf( __( 'You need to  provide Google Maps API license ( server key ) above. See more details in %shelp tab%s.', ud_get_wp_property('domain') ), '<a href="#tab-link-google-map-api-key" class="open-help-tab">', '</a>' ); ?></span>
           </li>
         </ul>
       </td>
@@ -266,6 +275,8 @@ if ( get_option( 'permalink_structure' ) == '' ) {
             <li>
               <?php echo WPP_F::checkbox( "name=wpp_settings[configuration][show_advanced_options]&label=" . __( 'Enable Standard Attributes Matching', ud_get_wp_property()->domain ), ( isset( $wp_properties[ 'configuration' ][ 'show_advanced_options' ] ) ? $wp_properties[ 'configuration' ][ 'show_advanced_options' ] : false ) ); ?>
               <i class="description wpp-notice-for-match" title="<?php _e( 'This option is designed to help us find which attribute you want to show as Price, Address, etc and place it in correct place in our templates.', ud_get_wp_property()->domain ); ?>"> ? </i>
+              <?php echo WPP_F::checkbox( "name=wpp_settings[configuration][pre_release_update]&label=" . __( 'Enable pre-release updates.', ud_get_wp_property()->domain ), ( isset( $wp_properties[ 'configuration' ][ 'pre_release_update' ] ) ? $wp_properties[ 'configuration' ][ 'pre_release_update' ] : false ) ); ?>
+              <br/>
             </li>
           </ul>
         </div>
@@ -555,6 +566,13 @@ if ( get_option( 'permalink_structure' ) == '' ) {
         <input type="button" class="button" value="<?php _e( 'Clear Cache', ud_get_wp_property()->domain ) ?>" id="wpp_clear_cache">
       </div>
 
+    <?php if(function_exists('icl_object_id')):?>
+      <div class="wpp_settings_block">
+        <?php _e( 'Generate images for duplicates of properties (WPML plugin option). ', ud_get_wp_property()->domain ) ?>
+        <input type="button" class="button" value="<?php _e( 'Generate', ud_get_wp_property()->domain ) ?>" id="wpp_is_remote_meta">
+      </div>
+    <?php endif;?>
+    
       <div class="wpp_settings_block"><?php printf(__( 'Set all %1s to same %2s type:', ud_get_wp_property()->domain ), WPP_F::property_label( 'plural' ), WPP_F::property_label( 'singular' )) ?>
         <select id="wpp_ajax_max_set_property_type_type">
         <?php foreach ( $wp_properties[ 'property_types' ] as $p_slug => $p_label ) { ?>